+++ docs/releases.md
@@ -10,6 +10,7 @@ uid: releases
 
 * Allow parallel processing of MQTT messages
 * Upgrade to [MQTTnet 4.2.1.781](https://github.com/chkr1011/MQTTnet/releases/tag/v4.2.1.781)
+* Upgrade to [Confluent.Kafka 2.2.0](https://github.com/confluentinc/confluent-kafka-dotnet/releases/tag/v2.2.0)
 
 ## [4.3.2](https://github.com/BEagle1984/silverback/releases/tag/v4.3.2)